+++ CoreGraphFramework/src/au/gov/asd/tac/constellation/graph/StoreGraph.java
@@ -19,7 +19,6 @@ import static au.gov.asd.tac.constellation.graph.GraphConstants.NOT_FOUND;
 import au.gov.asd.tac.constellation.graph.NativeAttributeType.NativeValue;
 import au.gov.asd.tac.constellation.graph.attribute.AttributeDescription;
 import au.gov.asd.tac.constellation.graph.attribute.AttributeRegistry;
-import au.gov.asd.tac.constellation.graph.attribute.IntegerAttributeDescription;
 import au.gov.asd.tac.constellation.graph.locking.GraphOperationMode;
 import au.gov.asd.tac.constellation.graph.locking.LockingTarget;
 import au.gov.asd.tac.constellation.graph.locking.ParameterReadAccess;
@@ -126,7 +125,7 @@ public class StoreGraph extends LockingTarget implements GraphWriteMethods, Seri
     private GraphEdit graphEdit;
     private int layerMaskSelectedAttributeId = Graph.NOT_FOUND;
     private int vertexLayerMaskAttribureId = Graph.NOT_FOUND;
-    private int transactionFilterBitmaskAttrId = Graph.NOT_FOUND;
+    private int transactionLayerMaskAttrId = Graph.NOT_FOUND;
     private int vertexLayerVisibilityAttributeId = Graph.NOT_FOUND;
     private int transactionLayerVisibilityAttributeId = Graph.NOT_FOUND;
     private int cameraAttributeId = Graph.NOT_FOUND;
@@ -195,9 +194,9 @@ public class StoreGraph extends LockingTarget implements GraphWriteMethods, Seri
                 // Loop through all transactions and remove bitmasks
                 final int transactionCount = getTransactionCount();
                 for (int i = 0; i < transactionCount; i++) {
-                    int bitmask = getIntValue(transactionFilterBitmaskAttrId, i);
+                    int bitmask = getIntValue(transactionLayerMaskAttrId, i);
                     bitmask = bitmask & ~(1 << count); // set bit to false
-                    setIntValue(transactionFilterBitmaskAttrId, i, bitmask);
+                    setIntValue(transactionLayerMaskAttrId, i, bitmask);
                 }
             }
             count++;
@@ -456,7 +455,7 @@ public class StoreGraph extends LockingTarget implements GraphWriteMethods, Seri
         // New Parameters copied
         this.layerMaskSelectedAttributeId = original.layerMaskSelectedAttributeId;
         this.vertexLayerMaskAttribureId = original.vertexLayerMaskAttribureId;
-        this.transactionFilterBitmaskAttrId = original.transactionFilterBitmaskAttrId;
+        this.transactionLayerMaskAttrId = original.transactionLayerMaskAttrId;
         this.vertexLayerVisibilityAttributeId = original.vertexLayerVisibilityAttributeId;
         this.transactionLayerVisibilityAttributeId = original.transactionLayerVisibilityAttributeId;
         this.transactionXAttributeId = original.transactionXAttributeId;
@@ -1734,7 +1733,7 @@ public class StoreGraph extends LockingTarget implements GraphWriteMethods, Seri
         layerMaskSelectedAttributeId = LAYER_MASK_SELECTED_ATTRIBUTE_LABEL.equals(label) ? attributeId : layerMaskSelectedAttributeId;
         vertexLayerMaskAttribureId = LAYER_MASK_ATTRIBUTE_LABEL.equals(label) ? (elementType == GraphElementType.VERTEX ? attributeId : vertexLayerMaskAttribureId) : vertexLayerMaskAttribureId;
         vertexLayerVisibilityAttributeId = LAYER_VISIBILITY_ATTRIBUTE_LABEL.equals(label) ? (elementType == GraphElementType.VERTEX ? attributeId : vertexLayerVisibilityAttributeId) : vertexLayerVisibilityAttributeId;
-        transactionFilterBitmaskAttrId = LAYER_MASK_ATTRIBUTE_LABEL.equals(label) ? (elementType == GraphElementType.TRANSACTION ? attributeId : transactionFilterBitmaskAttrId) : transactionFilterBitmaskAttrId;
+        transactionLayerMaskAttrId = LAYER_MASK_ATTRIBUTE_LABEL.equals(label) ? (elementType == GraphElementType.TRANSACTION ? attributeId : transactionLayerMaskAttrId) : transactionLayerMaskAttrId;
         transactionLayerVisibilityAttributeId = LAYER_VISIBILITY_ATTRIBUTE_LABEL.equals(label) ? (elementType == GraphElementType.TRANSACTION ? attributeId : transactionLayerVisibilityAttributeId) : transactionLayerVisibilityAttributeId;
 
         return attributeId;
@@ -1989,9 +1988,6 @@ public class StoreGraph extends LockingTarget implements GraphWriteMethods, Seri
     }
 
     /**
-     * TODO: The definitions outlined within this method can be refactored into
-     * a Concept. The attributes are specific to the layers implementation.
-     *
      * Loop over all queries, recalculate dynamic ones. Append bit flag to
      * bitmask when it satisfies the query.
      *
@@ -2003,20 +1999,14 @@ public class StoreGraph extends LockingTarget implements GraphWriteMethods, Seri
     private int recalculateLayerMask(final GraphElementType elementType, final int elementId) {
         int bitmask = 0;
         if (elementType == GraphElementType.VERTEX && vertexLayerMaskAttribureId != Graph.NOT_FOUND && vertexLayerVisibilityAttributeId != Graph.NOT_FOUND) {
-            vertexLayerMaskAttribureId = ensureAttribute(vertexLayerMaskAttribureId,
-                    GraphElementType.VERTEX, IntegerAttributeDescription.ATTRIBUTE_NAME,
-                    LAYER_MASK_ATTRIBUTE_LABEL, "Bitmask identifying the layers "
-                    + "this vertex belongs to", 1, null);
+            vertexLayerMaskAttribureId = LayersConcept.VertexAttribute.LAYER_MASK.get(this);
             if (vertexLayerMaskAttribureId != Graph.NOT_FOUND) {
                 bitmask = getIntValue(vertexLayerMaskAttribureId, elementId);
             }
-        } else if (elementType == GraphElementType.TRANSACTION && transactionFilterBitmaskAttrId != Graph.NOT_FOUND && transactionLayerVisibilityAttributeId != Graph.NOT_FOUND) {
-            transactionFilterBitmaskAttrId = ensureAttribute(transactionFilterBitmaskAttrId,
-                    GraphElementType.TRANSACTION, IntegerAttributeDescription.ATTRIBUTE_NAME,
-                    LAYER_MASK_ATTRIBUTE_LABEL, "Bitmask identifying the layers "
-                    + "this transaction belongs to", 1, null);
-            if (transactionFilterBitmaskAttrId != Graph.NOT_FOUND) {
-                bitmask = getIntValue(transactionFilterBitmaskAttrId, elementId);
+        } else if (elementType == GraphElementType.TRANSACTION && transactionLayerMaskAttrId != Graph.NOT_FOUND && transactionLayerVisibilityAttributeId != Graph.NOT_FOUND) {
+            transactionLayerMaskAttrId = LayersConcept.TransactionAttribute.LAYER_MASK.get(this);
+            if (transactionLayerMaskAttrId != Graph.NOT_FOUND) {
+                bitmask = getIntValue(transactionLayerMaskAttrId, elementId);
             }
         }
 
@@ -2151,158 +2141,106 @@ public class StoreGraph extends LockingTarget implements GraphWriteMethods, Seri
 
     // update the layer visibility of an element
     private void updateLayerMask(final int attributeId, final GraphElementType elementType, final int elementId, final int selectedLayerMask) {
-//        final int bitmask = recalculateLayerMask(elementType, elementId);
-//        transactionLayerVisibilityAttributeId = ensureAttribute(transactionLayerVisibilityAttributeId,
-//                GraphElementType.TRANSACTION,
-//                FloatAttributeDescription.ATTRIBUTE_NAME,
-//                LAYER_VISIBILITY_ATTRIBUTE_LABEL,
-//                "The visibility of the transaction given "
-//                + "the layers it belongs to", 1.0f, null);
-//
-//        vertexLayerVisibilityAttributeId = ensureAttribute(vertexLayerVisibilityAttributeId,
-//                GraphElementType.VERTEX,
-//                FloatAttributeDescription.ATTRIBUTE_NAME,
-//                LAYER_VISIBILITY_ATTRIBUTE_LABEL,
-//                "The visibility of the vertex given "
-//                + "the layers it belongs to", 1.0f, null);
-//
-//        if (elementType == GraphElementType.VERTEX) {
-//
-//            // when attr ids are found
-//            if (vertexLayerMaskAttribureId != Graph.NOT_FOUND && vertexLayerVisibilityAttributeId != Graph.NOT_FOUND) {
-//                // Attributes exist, update bitmask (this eventually becomes more comples to cover all layers). If the bitmask
-//                // is non-zero value then the element can be displayed, if not, layer filters will result in it being hidden.
-//
-//                if (attributeId != vertexLayerMaskAttribureId) { // set filter bitmask when not updating (prevents infinite loop)
-//                    avoidLayerUpdate = true; // cancels out the re-update when setting the int vals
-//                    setIntValue(vertexLayerMaskAttribureId, elementId, bitmask);
-//                    avoidLayerUpdate = false;
-//                }
-//                final float existingVisibility = getFloatValue(vertexLayerVisibilityAttributeId, elementId);
-//                if ((bitmask & currentVisibleMask) > 0) {
-//                    // A value > 0 indicates the object is mapped to at least one visible layer
-//                    if (existingVisibility != 1.0f) {
-//                        attributeDescriptions[vertexLayerVisibilityAttributeId].setFloat(elementId, 1.0f);
-//                        attributeIndices[vertexLayerVisibilityAttributeId].updateElement(elementId);
-//                        attributeModificationCounters[vertexLayerVisibilityAttributeId] += operationMode.getModificationIncrement();
-//                    }
-//                } else {
-//                    // A value = 0 indicates the object is not mapped to at least one visible layer
-//                    if (existingVisibility != 0.0f) {
-//                        attributeDescriptions[vertexLayerVisibilityAttributeId].setFloat(elementId, 0.0f);
-//                        attributeIndices[vertexLayerVisibilityAttributeId].updateElement(elementId);
-//                        attributeModificationCounters[vertexLayerVisibilityAttributeId] += operationMode.getModificationIncrement();
-//                    }
-//                }
-//            }
-//        } else if (elementType == GraphElementType.TRANSACTION
-//                && transactionFilterBitmaskAttrId != Graph.NOT_FOUND
-//                && transactionLayerVisibilityAttributeId != Graph.NOT_FOUND) {
-//            if (attributeId != transactionFilterBitmaskAttrId) {
-//                avoidLayerUpdate = true;
-//                setIntValue(transactionFilterBitmaskAttrId, elementId, bitmask);
-//                avoidLayerUpdate = false;
-//            }
-//            final float existingVisibility = getFloatValue(transactionLayerVisibilityAttributeId, elementId);
-//            if ((bitmask & currentVisibleMask) > 0) {
-//                if (existingVisibility != 1.0f) {
-//                    attributeDescriptions[transactionLayerVisibilityAttributeId].setFloat(elementId, 1.0f);
-//                    attributeIndices[transactionLayerVisibilityAttributeId].updateElement(elementId);
-//                    attributeModificationCounters[transactionLayerVisibilityAttributeId] += operationMode.getModificationIncrement();
-//                }
-//            } else {
-//                if (existingVisibility != 0.0f) {
-//                    attributeDescriptions[transactionLayerVisibilityAttributeId].setFloat(elementId, 0.0f);
-//                    attributeIndices[transactionLayerVisibilityAttributeId].updateElement(elementId);
-//                    attributeModificationCounters[transactionLayerVisibilityAttributeId] += operationMode.getModificationIncrement();
-//                }
-//            }
-//        }
+        final int bitmask = recalculateLayerMask(elementType, elementId);
+        transactionLayerVisibilityAttributeId = LayersConcept.TransactionAttribute.LAYER_VISIBILITY.get(this);
+        vertexLayerVisibilityAttributeId = LayersConcept.VertexAttribute.LAYER_VISIBILITY.get(this);
+
+        if (elementType == GraphElementType.VERTEX) {
+            // when attr ids are found
+            if (vertexLayerMaskAttribureId != Graph.NOT_FOUND && vertexLayerVisibilityAttributeId != Graph.NOT_FOUND) {
+                // Attributes exist, update bitmask (this eventually becomes more comples to cover all layers). If the bitmask
+                // is non-zero value then the element can be displayed, if not, layer filters will result in it being hidden.
+
+                if (attributeId != vertexLayerMaskAttribureId) { // set filter bitmask when not updating (prevents infinite loop)
+                    avoidLayerUpdate = true; // cancels out the re-update when setting the int vals
+                    setIntValue(vertexLayerMaskAttribureId, elementId, bitmask);
+                    avoidLayerUpdate = false;
+                }
+                final float existingVisibility = getFloatValue(vertexLayerVisibilityAttributeId, elementId);
+                if ((bitmask & currentVisibleMask) > 0) {
+                    // A value > 0 indicates the object is mapped to at least one visible layer
+                    if (existingVisibility != 1.0f) {
+                        attributeDescriptions[vertexLayerVisibilityAttributeId].setFloat(elementId, 1.0f);
+                        attributeIndices[vertexLayerVisibilityAttributeId].updateElement(elementId);
+                        attributeModificationCounters[vertexLayerVisibilityAttributeId] += operationMode.getModificationIncrement();
+                    }
+                } else {
+                    // A value = 0 indicates the object is not mapped to at least one visible layer
+                    if (existingVisibility != 0.0f) {
+                        attributeDescriptions[vertexLayerVisibilityAttributeId].setFloat(elementId, 0.0f);
+                        attributeIndices[vertexLayerVisibilityAttributeId].updateElement(elementId);
+                        attributeModificationCounters[vertexLayerVisibilityAttributeId] += operationMode.getModificationIncrement();
+                    }
+                }
+            }
+        } else if (elementType == GraphElementType.TRANSACTION
+                && transactionLayerMaskAttrId != Graph.NOT_FOUND
+                && transactionLayerVisibilityAttributeId != Graph.NOT_FOUND) {
+            if (attributeId != transactionLayerMaskAttrId) {
+                avoidLayerUpdate = true;
+                setIntValue(transactionLayerMaskAttrId, elementId, bitmask);
+                avoidLayerUpdate = false;
+            }
+            final float existingVisibility = getFloatValue(transactionLayerVisibilityAttributeId, elementId);
+            if ((bitmask & currentVisibleMask) > 0) {
+                if (existingVisibility != 1.0f) {
+                    attributeDescriptions[transactionLayerVisibilityAttributeId].setFloat(elementId, 1.0f);
+                    attributeIndices[transactionLayerVisibilityAttributeId].updateElement(elementId);
+                    attributeModificationCounters[transactionLayerVisibilityAttributeId] += operationMode.getModificationIncrement();
+                }
+            } else {
+                if (existingVisibility != 0.0f) {
+                    attributeDescriptions[transactionLayerVisibilityAttributeId].setFloat(elementId, 0.0f);
+                    attributeIndices[transactionLayerVisibilityAttributeId].updateElement(elementId);
+                    attributeModificationCounters[transactionLayerVisibilityAttributeId] += operationMode.getModificationIncrement();
+                }
+            }
+        }
     }
 
     // update the layer visibility for an element when an attribute changes
     private void updateLayerMask(final int attributeId, final int elementid) {
-//        if (transactionLayerVisibilityAttributeId == Graph.NOT_FOUND) {
-//            transactionLayerVisibilityAttributeId = this.addAttribute(
-//                    GraphElementType.TRANSACTION,
-//                    FloatAttributeDescription.ATTRIBUTE_NAME,
-//                    LAYER_VISIBILITY_ATTRIBUTE_LABEL,
-//                    "The visibility of the transaction given "
-//                    + "the layers it belongs to", 1.0f, null);
-//        }
-//        if (vertexLayerVisibilityAttributeId == Graph.NOT_FOUND) {
-//            vertexLayerVisibilityAttributeId = this.addAttribute(
-//                    GraphElementType.VERTEX,
-//                    FloatAttributeDescription.ATTRIBUTE_NAME,
-//                    LAYER_VISIBILITY_ATTRIBUTE_LABEL,
-//                    "The visibility of the vertex given "
-//                    + "the layers it belongs to", 1.0f, null);
-//        }
-//        if (layerMaskSelectedAttributeId == Graph.NOT_FOUND) {
-//            layerMaskSelectedAttributeId = this.addAttribute(
-//                    GraphElementType.GRAPH,
-//                    IntegerAttributeDescription.ATTRIBUTE_NAME,
-//                    LAYER_MASK_SELECTED_ATTRIBUTE_LABEL,
-//                    "The layers currently enabled for display", 1, null);
-//        }
-//        if (vertexLayerMaskAttribureId == Graph.NOT_FOUND) {
-//            vertexLayerMaskAttribureId = this.addAttribute(
-//                    GraphElementType.VERTEX,
-//                    IntegerAttributeDescription.ATTRIBUTE_NAME,
-//                    LAYER_MASK_ATTRIBUTE_LABEL,
-//                    "Bitmask identifying the layers this vertex belongs to", 1, null);
-//        }
-//        if (transactionFilterBitmaskAttrId == Graph.NOT_FOUND) {
-//            transactionFilterBitmaskAttrId = this.addAttribute(
-//                    GraphElementType.TRANSACTION,
-//                    IntegerAttributeDescription.ATTRIBUTE_NAME,
-//                    LAYER_MASK_ATTRIBUTE_LABEL,
-//                    "Bitmask identifying the layers this transaction belongs to", 1, null);
-//        }
-//        if (attributeId != vertexLayerVisibilityAttributeId && attributeId != transactionLayerVisibilityAttributeId) {
-//            // get the element type of selected layer mask and recalculate element mask
-//            currentVisibleMask = (layerMaskSelectedAttributeId != Graph.NOT_FOUND)
-//                    ? getIntValue(layerMaskSelectedAttributeId, 0) : 1;
-//            final GraphElementType elementType = getAttributeElementType(attributeId);
-//
-//            // preload layerPrefs jit if it hasn't been instantiated
-//            if (layerPrefs.isEmpty()) {
-//                for (int i = 0; i < 31; i++) {
-//                    layerPrefs.add((byte) 0b0);
-//                }
-//            }
-//
-//            updateLayerMask(attributeId, elementType, elementid, currentVisibleMask);
-//        }
+        layerMaskSelectedAttributeId = LayersConcept.GraphAttribute.LAYER_MASK_SELECTED.get(this);
+        vertexLayerMaskAttribureId = LayersConcept.VertexAttribute.LAYER_MASK.get(this);
+        vertexLayerVisibilityAttributeId = LayersConcept.VertexAttribute.LAYER_VISIBILITY.get(this);
+        transactionLayerMaskAttrId = LayersConcept.TransactionAttribute.LAYER_MASK.get(this);
+        transactionLayerVisibilityAttributeId = LayersConcept.TransactionAttribute.LAYER_VISIBILITY.get(this);
+
+        // TODO: not sure about why we have to keep id's recorded
+        if (attributeId != vertexLayerVisibilityAttributeId && attributeId != transactionLayerVisibilityAttributeId) {
+            // get the element type of selected layer mask and recalculate element mask
+            currentVisibleMask = (layerMaskSelectedAttributeId != Graph.NOT_FOUND)
+                    ? getIntValue(layerMaskSelectedAttributeId, 0) : 1;
+            final GraphElementType elementType = getAttributeElementType(attributeId);
+
+            // preload layerPrefs jit if it hasn't been instantiated
+            if (layerPrefs.isEmpty()) {
+                for (int i = 0; i < 31; i++) {
+                    layerPrefs.add((byte) 0b0);
+                }
+            }
+
+            updateLayerMask(attributeId, elementType, elementid, currentVisibleMask);
+        }
     }
 
     // update the layer visibility of all elements
     private void updateAllLayerMasks() {
-//        currentVisibleMask = (layerMaskSelectedAttributeId != Graph.NOT_FOUND)
-//                ? getIntValue(layerMaskSelectedAttributeId, 0) : 1;
-//
-//        recalculateLayerVisibilities();
-//
-//        // loop through all vertexes and recalculate masks
-//        final int vertexCount = getVertexCount();
-//        for (int i = 0; i < vertexCount; i++) {
-//            updateLayerMask(-1, GraphElementType.VERTEX, vStore.getElement(i), currentVisibleMask);
-//        }
-//
-//        // loop through all transactions and recalculate masks
-//        final int transactionCount = getTransactionCount();
-//        for (int i = 0; i < transactionCount; i++) {
-//            updateLayerMask(-1, GraphElementType.TRANSACTION, tStore.getElement(i), currentVisibleMask);
-//        }
-    }
+        currentVisibleMask = (layerMaskSelectedAttributeId != Graph.NOT_FOUND)
+                ? getIntValue(layerMaskSelectedAttributeId, 0) : 1;
 
-    private int ensureAttribute(final int attributeId, final GraphElementType elementType, final String attributeType, final String label,
-            final String description, final Object defaultValue, final String attributeMergerId) {
-        if (attributeDescriptions[attributeId] == null) {
-            return this.addAttribute(elementType, attributeType, label, description,
-                    defaultValue, attributeMergerId);
-        } else {
-            return this.getAttribute(elementType, label);
+        recalculateLayerVisibilities();
+
+        // loop through all vertexes and recalculate masks
+        final int vertexCount = getVertexCount();
+        for (int i = 0; i < vertexCount; i++) {
+            updateLayerMask(-1, GraphElementType.VERTEX, vStore.getElement(i), currentVisibleMask);
+        }
+
+        // loop through all transactions and recalculate masks
+        final int transactionCount = getTransactionCount();
+        for (int i = 0; i < transactionCount; i++) {
+            updateLayerMask(-1, GraphElementType.TRANSACTION, tStore.getElement(i), currentVisibleMask);
         }
     }