+++ CoreInteractiveGraph/src/au/gov/asd/tac/constellation/graph/interaction/plugins/draw/CreateTransactionPlugin.java
@@ -15,8 +15,9 @@
  */
 package au.gov.asd.tac.constellation.graph.interaction.plugins.draw;
 
+import au.gov.asd.tac.constellation.graph.Graph;
 import au.gov.asd.tac.constellation.graph.GraphWriteMethods;
-import au.gov.asd.tac.constellation.graph.schema.visual.concept.VisualConcept;
+import au.gov.asd.tac.constellation.graph.LayersConcept;
 import au.gov.asd.tac.constellation.plugins.Plugin;
 import au.gov.asd.tac.constellation.plugins.PluginInteraction;
 import au.gov.asd.tac.constellation.plugins.parameters.PluginParameter;
@@ -77,13 +78,17 @@ public final class CreateTransactionPlugin extends SimpleEditPlugin {
         this.destination = parameters.getParameters().get(DESTINATION_PARAMETER_ID).getIntegerValue();
         this.directed = parameters.getParameters().get(DIRECTED_PARAMETER_ID).getBooleanValue();
 
-        final int txLayerAttrId = VisualConcept.TransactionAttribute.LAYER_MASK.ensure(graph);
-        final int graphLayerAttrId = VisualConcept.GraphAttribute.LAYER_MASK_SELECTED.ensure(graph);
+        final int txLayerAttrId = LayersConcept.TransactionAttribute.LAYER_MASK.get(graph);
+        final int graphLayerAttrId = LayersConcept.GraphAttribute.LAYER_MASK_SELECTED.get(graph);
 
         final int txId = graph.addTransaction(source, destination, directed);
-        int layer = graph.getIntValue(graphLayerAttrId, 0);
-        layer = layer == 1 ? 1 : layer | (1 << 0);
-        graph.setIntValue(txLayerAttrId, txId, layer);
+        
+        // add layer mask attributes
+        if (graphLayerAttrId != Graph.NOT_FOUND && txLayerAttrId != Graph.NOT_FOUND) {
+            int layer = graph.getIntValue(graphLayerAttrId, 0);
+            layer = layer == 1 ? 1 : layer | (1 << 0);
+            graph.setIntValue(txLayerAttrId, txId, layer);
+        }
         graph.getSchema().newTransaction(graph, txId);
     }
 }