+++ CoreVisualGraph/src/au/gov/asd/tac/constellation/graph/visual/framework/GraphVisualAccess.java
@@ -18,6 +18,7 @@ package au.gov.asd.tac.constellation.graph.visual.framework;
 import au.gov.asd.tac.constellation.graph.Graph;
 import au.gov.asd.tac.constellation.graph.GraphElementType;
 import au.gov.asd.tac.constellation.graph.GraphReadMethods;
+import au.gov.asd.tac.constellation.graph.LayersConcept;
 import au.gov.asd.tac.constellation.graph.ReadableGraph;
 import au.gov.asd.tac.constellation.graph.schema.attribute.SchemaAttribute;
 import au.gov.asd.tac.constellation.graph.schema.visual.GraphLabel;
@@ -519,12 +520,10 @@ public final class GraphVisualAccess implements VisualAccess {
                 if (!Objects.equals(count, modCounts.put(VisualConcept.VertexAttribute.VISIBILITY, count))) {
                     changes.add(new VisualChangeBuilder(VisualProperty.VERTEX_VISIBILITY).forItems(accessGraph.getVertexCount()).build());
                 }
-                // TODO Start
                 count = vertexLayerVisibility == Graph.NOT_FOUND ? -1 : accessGraph.getValueModificationCounter(vertexLayerVisibility);
-                if (!Objects.equals(count, modCounts.put(VisualConcept.VertexAttribute.LAYER_VISIBILITY, count))) {
+                if (!Objects.equals(count, modCounts.put(LayersConcept.VertexAttribute.LAYER_VISIBILITY, count))) {
                     changes.add(new VisualChangeBuilder(VisualProperty.VERTEX_VISIBILITY).forItems(accessGraph.getVertexCount()).build());
                 }
-                // TODO endXXXX
                 count = vertexDimmed == Graph.NOT_FOUND ? -1 : accessGraph.getValueModificationCounter(vertexDimmed);
                 if (!Objects.equals(count, modCounts.put(VisualConcept.VertexAttribute.DIMMED, count))) {
                     changes.add(new VisualChangeBuilder(VisualProperty.VERTEX_DIM).forItems(accessGraph.getVertexCount()).build());
@@ -554,7 +553,7 @@ public final class GraphVisualAccess implements VisualAccess {
                     changes.add(new VisualChangeBuilder(VisualProperty.CONNECTION_VISIBILITY).forItems(connectionElementTypes.length).build());
                 }
                 count = transactionLayerVisibility == Graph.NOT_FOUND ? -1 : accessGraph.getValueModificationCounter(transactionLayerVisibility);
-                if (!Objects.equals(count, modCounts.put(VisualConcept.TransactionAttribute.LAYER_VISIBILITY, count))) {
+                if (!Objects.equals(count, modCounts.put(LayersConcept.TransactionAttribute.LAYER_VISIBILITY, count))) {
                     changes.add(new VisualChangeBuilder(VisualProperty.CONNECTION_VISIBILITY).forItems(connectionElementTypes.length).build());
                 }
                 count = transactionDimmed == Graph.NOT_FOUND ? -1 : accessGraph.getValueModificationCounter(transactionDimmed);
@@ -604,7 +603,7 @@ public final class GraphVisualAccess implements VisualAccess {
         vertexForegroundIcon = VisualConcept.VertexAttribute.FOREGROUND_ICON.get(rg);
         vertexSelected = VisualConcept.VertexAttribute.SELECTED.get(rg);
         vertexVisibility = VisualConcept.VertexAttribute.VISIBILITY.get(rg);
-        vertexLayerVisibility = VisualConcept.VertexAttribute.LAYER_VISIBILITY.get(rg);
+        vertexLayerVisibility = LayersConcept.VertexAttribute.LAYER_VISIBILITY.get(rg);
         vertexDimmed = VisualConcept.VertexAttribute.DIMMED.get(rg);
         vertexRadius = VisualConcept.VertexAttribute.NODE_RADIUS.get(rg);
         vertexBlaze = VisualConcept.VertexAttribute.BLAZE.get(rg);
@@ -612,7 +611,7 @@ public final class GraphVisualAccess implements VisualAccess {
         transactionSelected = VisualConcept.TransactionAttribute.SELECTED.get(rg);
         transactionDirected = VisualConcept.TransactionAttribute.DIRECTED.get(rg);
         transactionVisibility = VisualConcept.TransactionAttribute.VISIBILITY.get(rg);
-        transactionLayerVisibility = VisualConcept.TransactionAttribute.LAYER_VISIBILITY.get(rg);
+        transactionLayerVisibility = LayersConcept.TransactionAttribute.LAYER_VISIBILITY.get(rg);
         transactionDimmed = VisualConcept.TransactionAttribute.DIMMED.get(rg);
         transactionLineStyle = VisualConcept.TransactionAttribute.LINE_STYLE.get(rg);
         transactionWidth = VisualConcept.TransactionAttribute.WIDTH.get(rg);